moved everything to the hand file
